Posible corrección de la función mix de moulinette

+++ ft_printf/deal_with_hex.c
@@ -6,7 +6,7 @@
 /*   By: nmorgado <nmorgado@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 09:09:27 by nmorgado          #+#    #+#             */
-/*   Updated: 2025/03/11 12:54:59 by nmorgado         ###   ########.fr       */
+/*   Updated: 2025/03/17 12:32:27 by nmorgado         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ int	deal_with_hex(va_list args, char **string, char format)
 	else
 		cap_bool = 1;
 	string2 = 0;
-	integer = va_arg(args, unsigned long int);
+	integer = va_arg(args, unsigned int);
 	str_int = ft_make_hex(integer, cap_bool);
 	if (str_int == 0)
 	{
